move globals to own file
