+++ test/src/globals.h
@@ -14,10 +14,15 @@
  * limitations under the License.
  */
 
+#pragma once
+
 #include "NTest.h"
+
 
 namespace Nakama {
     namespace Test {
+
+        #define NTEST_ASSERT(cond)  if (!(cond)) { abortCurrentTest(__FILE__, __LINE__); }
 
         // currently running tests
         extern std::vector<NTest*> g_running_tests;
@@ -27,5 +32,11 @@ namespace Nakama {
         extern uint32_t g_runTestsCount;
         extern uint32_t g_failedTestsCount;
 
+        void sleep(uint32_t ms);
+        int getFailedCount();
+        void abortCurrentTest(const char* file, int lineno);
+        void removeRunningTest(NTest* test);
+        void addRunningTest(NTest* test);
+        void runTestsLoop();
     }
 }
